+++ src/controlador/DAO/AlumnoDAO.java
@@ -30,7 +30,7 @@ public interface AlumnoDAO {
     List<Alumno> empiezaPorDni(String dni);
     List<Alumno> terminaEnDni(String dni);
 
-    List<Alumno> notaMediaAlum(int mediia);
+    List<Alumno> notaMediaAlum(Double mediia);
     List<Alumno> profesorTutorAlum(String nombreTutor);
 
     List<Alumno> buscarCursoAlum(List<String> listaCur);
